fix: move SearchSection component to SearchSection folder
