+++ src/components/SearchSection/SearchSection.tsx
@@ -1,6 +1,6 @@
 import { Component, createRef, type ReactNode } from 'react';
-import { SEARCH_VALUE } from '../constants/consts';
-import Button from './Button/Button';
+import { SEARCH_VALUE } from '../../constants/consts';
+import Button from '../Button/Button';
 
 interface Props {
   fetchData: (value: string | undefined) => void;
